+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changes to Joyride
 ## [Unreleased]
 
 - [Add `joyride.core/*file*`](https://github.com/BetterThanTomorrow/joyride/issues/5)
+- Change name of command *Run Script* -> *Run Clojure Code*
 
 ## [0.0.2] - 2022-04-27